pyproject.toml
[project]
name = "deformers"
version = "0.2.2"
description = "Reverse engineer transformer models."
authors = [{ name = "apehex", email = "17317183+apehex@users.noreply.github.com" }]
readme = ".github/README.md"
license = "AGPL-3.0"
license-files = [".github/LICENSE.md"]
requires-python = ">=3.10, <3.14"
dependencies = [
    "mlable-torch>=0.4",
    "transformers>=5.2",]

[dependency-groups]
dev = ["pytest"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/deformers"]

[tool.hatch.build.targets.sdist]
include = ["src/deformers", "pyproject.toml", ".github/README.md", ".github/LICENSE.md"]
